Graphical elements and slider each moved into a new class.
