+++ CartPole/__init__.py
@@ -16,6 +16,7 @@ import timeit
 
 import numpy as np
 import pandas as pd
+import matplotlib.pyplot as plt
 
 from Control_Toolkit.Controllers import template_controller
 from Control_Toolkit.others.environment import EnvironmentBatched
@@ -47,26 +48,18 @@ s0 = create_cartpole_state()
 # check memory usage of chosen methods. Commented by default
 # from memory_profiler import profile
 
-# region Graphics imports
-import matplotlib.pyplot as plt
-# rc sets global parameters for matplotlib; transforms is used to rotate the Mast
-from matplotlib import animation, rc, transforms
-# Shapes used to draw a Cart and the slider
-from matplotlib.patches import (Circle, FancyArrowPatch, FancyBboxPatch,
-                                Rectangle)
+
 
 from random import random
 
 # Angle convention to rotate the mast in right direction - depends on used Equation
-from CartPole.cartpole_equations import ANGLE_CONVENTION, CartPoleEquations
+from CartPole.cartpole_equations import CartPoleEquations
 
 from CartPole.csv_logger import create_csv_file_name, create_csv_title, create_csv_header, create_csv_file, save_data_to_csv_file
 
 from SI_Toolkit.Functions.FunctionalDict import FunctionalDict, HistoryClass
 
-# Set the font parameters for matplotlib figures
-font = {'size': 22}
-rc('font', **font)
+
 # endregion
 
 # endregion
@@ -81,9 +74,12 @@ class CartPole(EnvironmentBatched):
     num_states = 6
     num_acions = 1
 
-    def __init__(self, initial_state=s0, path_to_experiment_recordings=None):
+    def __init__(self, initial_state=s0, path_to_experiment_recordings=None, target_slider=None):
+
         self.config = config["cartpole"]
         self.rng_CartPole = create_rng(self.__class__.__name__, self.config["seed"])
+
+        self.slider = target_slider
 
         self.time_L_last_change = None
         self.start_changing_L = False
@@ -245,58 +241,6 @@ class CartPole(EnvironmentBatched):
 
         self.dict_history = HistoryClass()  # Class holding the history of the simulation
         self.dict_history.add_keys(keys=self.variables_to_log.keys())
-
-        # region Variables initialization for drawing/animating a CartPole
-        # DIMENSIONS OF THE DRAWING ONLY!!!
-        # NOTHING TO DO WITH THE SIMULATION AND NOT INTENDED TO BE MANIPULATED BY USER !!!
-
-        # Variable relevant for interactive use of slider
-        self.slider_max = 1.0
-        self.slider_value = 0.0
-
-        self.show_hanging_pole = False
-
-        self.physical_to_graphics = None
-        self.graphics_to_physical = None
-
-        # Parameters needed to display CartPole in GUI
-        # They are assigned with values in self.init_elements()
-        self.CartLength = None
-        self.WheelRadius = None
-        self.WheelToMiddle = None
-        self.y_plane = None
-        self.y_wheel = None
-        self.mast_height_maximal_drawing_units = None  # For drawing only. For calculation see L
-        self.max_height_maximal_physical_units = None
-        self.mast_height_current_drawing_units = None
-        self.zero_angle_tick_height_current_drawing_units = None
-        self.MastThickness = None
-        self.ZeroAngleTickThickness = None
-        self.TrackHalfLengthGraphics = None  # Length of the track
-
-        # Elements of the drawing
-        self.Mast = None
-        self.Chassis = None
-        self.WheelLeft = None
-        self.WheelRight = None
-
-        self.ZeroAngleTick = None
-
-        # Arrow indicating acceleration (=motor power)
-        self.Acceleration_Arrow = None
-
-        self.y_acceleration_arrow = None
-        self.scaling_dx_acceleration_arrow = None
-        self.x_acceleration_arrow = None
-
-        # Depending on mode, slider may be displayed either as bar or as an arrow
-        self.Slider_Bar = None
-        self.Slider_Arrow = None
-        self.t2 = None  # An abstract container for the transform rotating the mast
-        self.t_zero_angle = None  # An abstract container for the transform rotating the zero angle tick
-
-        self.init_graphical_elements()  # Assign proper object to the above variables
-        # endregion
         
         self.target_position = 0.0
         self.target_equilibrium = 1.0
@@ -411,9 +355,9 @@ class CartPole(EnvironmentBatched):
                     return
 
                 self.target_position = self.random_track_f(self.time)
-                self.slider_value = self.target_position/TrackHalfLength  # Assign target position to slider to display it
+                self.slider.value = self.target_position/TrackHalfLength  # Assign target position to slider to display it
             else:
-                self.target_position = self.slider_value * TrackHalfLength  # Get target position from slider
+                self.target_position = self.slider.value * TrackHalfLength  # Get target position from slider
 
     def update_target_equilibrium(self):
         if self.time_last_target_equilibrium_change is None:
@@ -520,7 +464,7 @@ class CartPole(EnvironmentBatched):
 
             if self.controller_name == 'manual-stabilization':
                 # in this case slider corresponds already to the power of the motor
-                self.Q_calculated = self.slider_value
+                self.Q_calculated = self.slider.value
                 self.Q_update_time = 0.0
             else:  # in this case slider gives a target position, lqr regulator
                 update_start = timeit.default_timer()
@@ -876,9 +820,9 @@ class CartPole(EnvironmentBatched):
                 
         # Set the maximal allowed value of the slider - relevant only for GUI
         if self.controller_name == 'manual-stabilization':
-            self.Slider_Arrow.set_positions((0, 0), (0, 0))
+            self.slider.Slider_Arrow.set_positions((0, 0), (0, 0))
         else:
-            self.Slider_Bar.set_width(0.0)
+            self.slider.Slider_Bar.set_width(0.0)
 
         # TODO: optimally reset_dict_history would be False and the controller could be switched during experiment
         #   The False option is not implemented yet. So it is possible to switch controller only when the experiment is not running.
@@ -917,7 +861,7 @@ class CartPole(EnvironmentBatched):
             self.s[ANGLE_SIN_IDX] = np.sin(self.s[ANGLE_IDX])
 
             self.target_position = 0.0
-            self.slider_value = 0.0
+            self.slider.value = 0.0
 
         elif reset_mode == 1:  # You may change this but be careful with other user. Better use 3
             # You can change here with which initial parameters you wish to start the simulation
@@ -930,7 +874,7 @@ class CartPole(EnvironmentBatched):
             self.s[ANGLE_SIN_IDX] = np.sin(self.s[ANGLE_IDX])
 
             self.target_position = 0.0
-            self.slider_value = 0.0
+            self.slider.value = 0.0
 
         elif reset_mode == 2:  # Don't change it
             if (s is not None) and (target_position is not None):
@@ -939,14 +883,14 @@ class CartPole(EnvironmentBatched):
                 self.s[ANGLE_COS_IDX] = np.cos(self.s[ANGLE_IDX])
                 self.s[ANGLE_SIN_IDX] = np.sin(self.s[ANGLE_IDX])
 
-                self.slider = self.target_position = target_position
+                self.slider.value = self.target_position = target_position
 
             else:
                 raise ValueError('s, Q or target position not provided for initial state')
 
             if self.controller_name == 'manual-stabilization':
                 # in this case slider corresponds already to the power of the motor
-                self.Q_calculated = self.slider_value
+                self.Q_calculated = self.slider.value
             else:  # in this case slider gives a target position, lqr regulator
                 self.Q_calculated = float(self.controller.step(
                     self.s,
@@ -1039,255 +983,9 @@ class CartPole(EnvironmentBatched):
 
     # endregion
 
-    # region 5. Methods needed to display CartPole in GUI
-    """
-    This section contains methods related to displaying CartPole in GUI of the simulator.
-    One could think of moving these function outside of CartPole class and connecting them rather more tightly
-    with GUI of the simulator.
-    We leave them however as a part of CartPole class as they rely on variables of the CartPole.
-    """
-
-    # This method initializes CartPole elements to be plotted in CartPole GUI
-    def init_graphical_elements(self):
-
-        self.CartLength = 10.0
-        self.WheelRadius = 0.5
-        self.WheelToMiddle = 4.0
-        self.y_plane = 0.0
-        self.y_wheel = self.y_plane + self.WheelRadius
-
-        self.mast_height_maximal_drawing_units = 10.0
-        self.max_height_maximal_physical_units = np.max([self.L_initial, *self.L_range])
-        self.mast_height_current_drawing_units = self.mast_height_maximal_drawing_units * (float(L) / self.max_height_maximal_physical_units)
-
-        self.zero_angle_tick_height_current_drawing_units = 1.0
-
-        self.MastThickness = 0.05
-        self.ZeroAngleTickThickness = 0.01
-        self.TrackHalfLengthGraphics = 50.0  # Full Length of the track
-
-        self.physical_to_graphics = (self.TrackHalfLengthGraphics-self.WheelToMiddle)/TrackHalfLength  # TrackHalfLength is the effective length of track
-        self.graphics_to_physical = 1.0/self.physical_to_graphics
-
-        self.y_acceleration_arrow = 1.5 * self.WheelRadius
-        self.scaling_dx_acceleration_arrow = 20.0
-        self.x_acceleration_arrow = (
-                                   self.s[POSITION_IDX]*self.physical_to_graphics +
-                                   # np.sign(self.Q) * (self.CartLength / 2.0) +
-                                   self.scaling_dx_acceleration_arrow * self.Q
-        )
-
-        # Initialize elements of the drawing
-        self.Mast = FancyBboxPatch(xy=(self.s[POSITION_IDX]*self.physical_to_graphics - (self.MastThickness / 2.0), 1.25 * self.WheelRadius),
-                                   width=self.MastThickness,
-                                   height=self.mast_height_current_drawing_units,
-                                   fc='g')
-
-        self.ZeroAngleTick = FancyBboxPatch(xy=(self.s[POSITION_IDX]*self.physical_to_graphics - (self.ZeroAngleTickThickness / 2.0), 1.2 * self.mast_height_current_drawing_units),
-                                   width=self.ZeroAngleTickThickness,
-                                   height=self.zero_angle_tick_height_current_drawing_units,
-                                   fc='yellow')
-
-        self.Chassis = FancyBboxPatch((self.s[POSITION_IDX]*self.physical_to_graphics - (self.CartLength / 2.0), self.WheelRadius),
-                                      self.CartLength,
-                                      1 * self.WheelRadius,
-                                      fc='r')
-
-        self.WheelLeft = Circle((self.s[POSITION_IDX]*self.physical_to_graphics - self.WheelToMiddle, self.y_wheel),
-                                radius=self.WheelRadius,
-                                fc='y',
-                                ec='k',
-                                lw=5)
-
-        self.WheelRight = Circle((self.s[POSITION_IDX]*self.physical_to_graphics + self.WheelToMiddle, self.y_wheel),
-                                 radius=self.WheelRadius,
-                                 fc='y',
-                                 ec='k',
-                                 lw=5)
-
-        self.Acceleration_Arrow = FancyArrowPatch((self.s[POSITION_IDX]*self.physical_to_graphics, self.y_acceleration_arrow),
-                                                  (self.x_acceleration_arrow, self.y_acceleration_arrow),
-                                                  arrowstyle='simple', mutation_scale=10,
-                                                  facecolor='gold', edgecolor='orange')
-
-        self.Slider_Arrow = FancyArrowPatch((self.slider_value, 0), (self.slider_value, 0),
-                                            arrowstyle='fancy', mutation_scale=50)
-        self.Slider_Bar = Rectangle((0.0, 0.0), self.slider_value, 1.0)
-        self.t2 = transforms.Affine2D().rotate(0.0)  # An abstract container for the transform rotating the mast
-        self.t_zero_angle = transforms.Affine2D().rotate(0.0)  # An abstract container for the transform rotating the mast
-
-    # This method accepts the mouse position and updated the slider value accordingly
-    # The mouse position has to be captured by a function not included in this class
-    def update_slider(self, mouse_position):
-        # The if statement formulates a saturation condition
-
-        if mouse_position > self.slider_max:
-            self.slider_value = self.slider_max
-        elif mouse_position < -self.slider_max:
-            self.slider_value = -self.slider_max
-        else:
-            self.slider_value = mouse_position
-
-    # This method draws elements and set properties of the CartPole figure
-    # which do not change at every frame of the animation
-    def draw_constant_elements(self, fig, AxCart, AxSlider):
-
-        ## Upper chart with Cart Picture
-        # Set x and y limits
-        AxCart.set_xlim((-self.TrackHalfLengthGraphics * 1.1, self.TrackHalfLengthGraphics * 1.1))
-        AxCart.set_ylim((-1.0, 15.0))
-        # Remove ticks on the y-axes
-        AxCart.yaxis.set_major_locator(plt.NullLocator())  # NullLocator is used to disable ticks on the Figures
-
-        locs = [-50.0, -25.0, - 0.0, 25.0, 50.0]
-        labels = [str(np.around(np.array(x*self.graphics_to_physical), 3)) for x in locs]
-        AxCart.xaxis.set_major_locator(plt.FixedLocator(locs))
-        AxCart.xaxis.set_major_formatter(plt.FixedFormatter(labels))
-
-        # Draw track
-        Floor = Rectangle((-self.TrackHalfLengthGraphics, -1.0),
-                          2 * self.TrackHalfLengthGraphics,
-                          1.0,
-                          fc='brown')
-        AxCart.add_patch(Floor)
-
-        # Draw an invisible point at constant position
-        # Thanks to it the axes is drawn high enough for the mast
-        InvisiblePointUp = Rectangle((0, self.mast_height_maximal_drawing_units + 2.0),
-                                     self.MastThickness,
-                                     0.0001,
-                                     fc='w',
-                                     ec='w')
-
-        AxCart.add_patch(InvisiblePointUp)
-
-        if self.show_hanging_pole:
-            InvisiblePointDown = Rectangle((0, -self.mast_height_maximal_drawing_units - 2.0),
-                                         self.MastThickness,
-                                         0.0001,
-                                         fc='w',
-                                         ec='w')
-
-            AxCart.add_patch(InvisiblePointDown)
-
-        # Apply scaling
-        AxCart.axis('scaled')
-
-        ## Lower Chart with Slider
-        # Set y limits
-        AxSlider.set(xlim=(-1.1 * self.slider_max, self.slider_max * 1.1))
-        # Remove ticks on the y-axes
-        AxSlider.yaxis.set_major_locator(plt.NullLocator())  # NullLocator is used to disable ticks on the Figures
-
-        if self.controller_name == 'manual-stabilization':
-            pass
-        else:
-            locs = np.array([-50.0, -37.5, -25.0, -12.5, - 0.0, 12.5, 25.0, 37.5, 50.0])/50.0
-            labels = [str(np.around(np.array(x * TrackHalfLength), 3)) for x in locs]
-            AxSlider.xaxis.set_major_locator(plt.FixedLocator(locs))
-            AxSlider.xaxis.set_major_formatter(plt.FixedFormatter(labels))
-
-        # Apply scaling
-        AxSlider.set_aspect("auto")
-
-        return fig, AxCart, AxSlider
-
-    # This method updates the elements of the Cart Figure which change at every frame.
-    # Not that these elements are not ploted directly by this method
-    # but rather returned as objects which can be used by another function
-    # e.g. animation function from matplotlib package
-    def update_drawing(self):
-
-        self.x_acceleration_arrow = (
-                                   self.s[POSITION_IDX]*self.physical_to_graphics +
-                                   # np.sign(self.Q) * (self.CartLength / 2.0) +
-                                   self.scaling_dx_acceleration_arrow * self.Q
-        )
-
-        self.Acceleration_Arrow.set_positions((self.s[POSITION_IDX]*self.physical_to_graphics, self.y_acceleration_arrow),
-                                             (self.x_acceleration_arrow, self.y_acceleration_arrow))
-
-        # Draw mast
-        mast_position = (self.s[POSITION_IDX]*self.physical_to_graphics - (self.MastThickness / 2.0))
-        self.Mast.set_x(mast_position)
-        self.Mast.set_height(self.mast_height_maximal_drawing_units * (float(L) / self.max_height_maximal_physical_units))
-
-        zero_tick_position = (self.s[POSITION_IDX]*self.physical_to_graphics - (self.ZeroAngleTickThickness / 2.0))
-        self.ZeroAngleTick.set_x(zero_tick_position)
-
-        # Draw rotated mast
-        t21 = transforms.Affine2D().translate(-mast_position, -1.25 * self.WheelRadius)
-        if ANGLE_CONVENTION == 'CLOCK-NEG':
-            t22 = transforms.Affine2D().rotate(self.s[ANGLE_IDX])
-        elif ANGLE_CONVENTION == 'CLOCK-POS':
-            t22 = transforms.Affine2D().rotate(-self.s[ANGLE_IDX])
-        else:
-            raise ValueError('Unknown angle convention')
-        t23 = transforms.Affine2D().translate(mast_position, 1.25 * self.WheelRadius)
-        self.t2 = t21 + t22 + t23
-
-        t21 = transforms.Affine2D().translate(-zero_tick_position, 0.0)
-        if ANGLE_CONVENTION == 'CLOCK-NEG':
-            t22 = transforms.Affine2D().rotate(-self.zero_angle_shift)
-        elif ANGLE_CONVENTION == 'CLOCK-POS':
-            t22 = transforms.Affine2D().rotate(self.zero_angle_shift)
-        else:
-            raise ValueError('Unknown angle convention')
-        t23 = transforms.Affine2D().translate(zero_tick_position, 0.0)
-        self.t_zero_angle = t21 + t22 + t23
-
-        # Draw Chassis
-        self.Chassis.set_x(self.s[POSITION_IDX]*self.physical_to_graphics - (self.CartLength / 2.0))
-        # Draw Wheels
-        self.WheelLeft.center = (self.s[POSITION_IDX]*self.physical_to_graphics - self.WheelToMiddle, self.y_wheel)
-        self.WheelRight.center = (self.s[POSITION_IDX]*self.physical_to_graphics + self.WheelToMiddle, self.y_wheel)
-        # Draw SLider
-        if self.controller_name == 'manual-stabilization':
-            self.Slider_Bar.set_width(self.slider_value)
-        else:
-            self.Slider_Arrow.set_positions((self.slider_value, 0), (self.slider_value, 1.0))
-
-        return self.Mast, self.t2, self.Chassis, self.WheelRight, self.WheelLeft,\
-               self.Slider_Bar, self.Slider_Arrow, self.Acceleration_Arrow, self.ZeroAngleTick, self.t_zero_angle
-
-    # A function redrawing the changing elements of the Figure
-    def run_animation(self, fig):
-        def init():
-            # Adding variable elements to the Figure
-            fig.AxCart.add_patch(self.Mast)
-            fig.AxCart.add_patch(self.Chassis)
-            fig.AxCart.add_patch(self.WheelLeft)
-            fig.AxCart.add_patch(self.WheelRight)
-            fig.AxCart.add_patch(self.Acceleration_Arrow)
-            fig.AxCart.add_patch(self.ZeroAngleTick)
-            fig.AxSlider.add_patch(self.Slider_Bar)
-            fig.AxSlider.add_patch(self.Slider_Arrow)
-            return self.Mast, self.Chassis, self.WheelLeft, self.WheelRight,\
-                   self.Slider_Bar, self.Slider_Arrow, self.Acceleration_Arrow, self.ZeroAngleTick
-
-        def animationManage(i):
-            # Updating variable elements
-            self.update_drawing()
-            # Special care has to be taken of the mast rotation
-            self.t2 = self.t2 + fig.AxCart.transData
-            self.t_zero_angle = self.t_zero_angle + fig.AxCart.transData
-            self.Mast.set_transform(self.t2)
-            self.ZeroAngleTick.set_transform(self.t_zero_angle)
-            return self.Mast, self.Chassis, self.WheelLeft, self.WheelRight,\
-                   self.Slider_Bar, self.Slider_Arrow, self.Acceleration_Arrow, self.ZeroAngleTick
-
-        # Initialize animation object
-        anim = animation.FuncAnimation(fig, animationManage,
-                                       init_func=init,
-                                       frames=300,
-                                       # fargs=(CartPoleInstance,), # It was used when this function was a part of GUI class. Now left as an example how to add arguments to FuncAnimation
-                                       interval=10,
-                                       blit=True,
-                                       repeat=True)
-        return anim
-
-    # endregion
-
+    @staticmethod
+    def current_L():
+        return L
